_old_ref/requirements.txt
Django
python-dotenv
mammoth
matplotlib
seaborn
python-dateutil
numpy
pandas
scipy
pyodbc
supabase
djangorestframework==3.14.0
django-cors-headers==4.3.1
beautifulsoup4 
python-docx
pillow

anthropic
# google-generativeai
google-genai
langchain-google-genai
openai
langchain
langchain-anthropic

gunicorn
whitenoise
openpyxl
gotrue

oracledb
pandas
SQLAlchemy

jinja2